Face detection test: refactor stats, add few options, simplify and reformat code

+++ app/src/main/java/org/pw/engithesis/androidcameracontrol/detectors/eyepupildetectionalgorithms/EyePupilDetectionProjection.java
@@ -105,7 +105,7 @@ public class EyePupilDetectionProjection extends EyePupilDetectionAlgorithm {
         int index = startIndex;
 
         for (int i = 0; i < length; i++) {
-            sumColIntensity += (double) Utility.byteToInt(pixels[index]);
+            sumColIntensity += Utility.byteToInt(pixels[index]);
             index += step;
         }
